Ioi comments and moved tests
