+++ test/hotspot/gtest/utilities/test_bitMap.cpp
@@ -114,6 +114,133 @@ class BitMapTest {
 #endif
 };
 
+class BitMapTruncateTest {
+
+ public:
+  const static BitMap::idx_t BITMAP_SIZE = 128;
+
+  template <class ResizableBitMapClass>
+  static void fillBitMap(ResizableBitMapClass& map,  BitMap::idx_t size) {
+    BitMap::idx_t set_bits[] = {0, 31, 63, 64, 95, 127};
+    for (BitMap::idx_t bit : set_bits) {
+        if (bit < size) {
+            map.set_bit(bit);
+        }
+    }
+  }
+
+  template <class ResizableBitMapClass>
+  static void testTruncateOneWord() {
+    ResourceMark rm;
+
+    ResizableBitMapClass map(64);
+    map.set_bit(0);
+    map.set_bit(1);
+    map.set_bit(2);
+    map.set_bit(3);
+
+    ResizableBitMapClass result(2);
+    result.set_bit(0);
+    result.set_bit(1);
+
+    map.truncate(1, 3);
+
+    EXPECT_TRUE(map.is_same(result));
+  }
+
+  template <class ResizableBitMapClass>
+  static void testTruncateSame() {
+    // Resulting map should be the same as the original
+    ResourceMark rm;
+    ResizableBitMapClass expected(BITMAP_SIZE);
+    fillBitMap(expected, BITMAP_SIZE);
+
+    ResizableBitMapClass map(BITMAP_SIZE);
+    fillBitMap(map, BITMAP_SIZE);
+    map.truncate(0, BITMAP_SIZE);
+
+    EXPECT_TRUE(map.is_same(expected));
+  }
+
+  template <class ResizableBitMapClass>
+  static void testTruncateStart() {
+    // Resulting map should start at the beginning of the original
+    ResourceMark rm;
+    ResizableBitMapClass expected(64);
+    fillBitMap(expected, 64);
+
+    ResizableBitMapClass map(BITMAP_SIZE);
+    fillBitMap(map, BITMAP_SIZE);
+    map.truncate(0, 64);
+
+    EXPECT_TRUE(map.is_same(expected));
+  }
+
+  template <class ResizableBitMapClass>
+  static void testTruncateEnd() {
+    // Resulting map should end at the end of the original
+    ResourceMark rm;
+    ResizableBitMapClass expected(64);
+    expected.set_bit(0);
+    expected.set_bit(31);
+    expected.set_bit(63);
+
+    ResizableBitMapClass map(BITMAP_SIZE);
+    fillBitMap(map, BITMAP_SIZE);
+    map.truncate(64, 128);
+
+    EXPECT_TRUE(map.is_same(expected));
+  }
+
+  template <class ResizableBitMapClass>
+  static void testTruncateMiddle() {
+    // Resulting map should end at the end of the original
+    ResourceMark rm;
+    ResizableBitMapClass expected(64);
+    expected.set_bit(31);
+    expected.set_bit(32);
+    expected.set_bit(63);
+
+    ResizableBitMapClass map(BITMAP_SIZE);
+    fillBitMap(map, BITMAP_SIZE);
+    map.truncate(32, 96);
+
+    EXPECT_TRUE(map.is_same(expected));
+  }
+
+  template <class ResizableBitMapClass>
+  static void testTruncateStartUnaligned() {
+    // Resulting map should start at the beginning of the original
+    ResourceMark rm;
+    ResizableBitMapClass expected(96);
+    fillBitMap(expected, 96);
+
+    ResizableBitMapClass map(BITMAP_SIZE);
+    fillBitMap(map, BITMAP_SIZE);
+    map.truncate(0, 96);
+
+    EXPECT_TRUE(map.is_same(expected));
+  }
+
+  template <class ResizableBitMapClass>
+  static void testTruncateEndUnaligned() {
+    // Resulting map should end at the end of the original
+    ResourceMark rm;
+    ResizableBitMapClass expected(97);
+    expected.set_bit(0);
+    expected.set_bit(32);
+    expected.set_bit(33);
+    expected.set_bit(64);
+    expected.set_bit(96);
+
+    ResizableBitMapClass map(BITMAP_SIZE);
+    fillBitMap(map, BITMAP_SIZE);
+    map.truncate(31, 128);
+
+    EXPECT_TRUE(map.is_same(expected));
+  }
+};
+
 // TestArenaBitMap is the shorthand combination of Arena and ArenaBitMap.
 // Multiple inheritance guarantees to construct Arena first.
 class TestArenaBitMap : private Arena, public ArenaBitMap {
@@ -204,3 +331,66 @@ TEST_VM(BitMap, print_on) {
 }
 
 #endif
+
+TEST_VM(BitMap, truncate_same) {
+  BitMapTruncateTest::testTruncateSame<ResourceBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ResourceBitMap";
+  BitMapTruncateTest::testTruncateSame<TestCHeapBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type CHeapBitMap";
+  BitMapTruncateTest::testTruncateSame<TestArenaBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ArenaBitMap";
+}
+
+TEST_VM(BitMap, truncate_start) {
+  BitMapTruncateTest::testTruncateStart<ResourceBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ResourceBitMap";
+  BitMapTruncateTest::testTruncateStart<TestCHeapBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type CHeapBitMap";
+  BitMapTruncateTest::testTruncateStart<TestArenaBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ArenaBitMap";
+}
+
+TEST_VM(BitMap, truncate_end) {
+  BitMapTruncateTest::testTruncateEnd<ResourceBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ResourceBitMap";
+  BitMapTruncateTest::testTruncateEnd<TestCHeapBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type CHeapBitMap";
+  BitMapTruncateTest::testTruncateEnd<TestArenaBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ArenaBitMap";
+}
+
+TEST_VM(BitMap, truncate_middle) {
+  BitMapTruncateTest::testTruncateMiddle<ResourceBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ResourceBitMap";
+  BitMapTruncateTest::testTruncateMiddle<TestCHeapBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type CHeapBitMap";
+  BitMapTruncateTest::testTruncateMiddle<TestArenaBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ArenaBitMap";
+}
+
+TEST_VM(BitMap, truncate_start_unaligned) {
+  BitMapTruncateTest::testTruncateStartUnaligned<ResourceBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ResourceBitMap";
+  BitMapTruncateTest::testTruncateStartUnaligned<TestCHeapBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type CHeapBitMap";
+  BitMapTruncateTest::testTruncateStartUnaligned<TestArenaBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ArenaBitMap";
+}
+
+TEST_VM(BitMap, truncate_end_unaligned) {
+  BitMapTruncateTest::testTruncateEndUnaligned<ResourceBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ResourceBitMap";
+  BitMapTruncateTest::testTruncateEndUnaligned<TestCHeapBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type CHeapBitMap";
+  BitMapTruncateTest::testTruncateEndUnaligned<TestArenaBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ArenaBitMap";
+}
+
+TEST_VM(BitMap, truncate_one_word) {
+  BitMapTruncateTest::testTruncateOneWord<ResourceBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ResourceBitMap";
+  BitMapTruncateTest::testTruncateOneWord<TestCHeapBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type CHeapBitMap";
+  BitMapTruncateTest::testTruncateOneWord<TestArenaBitMap>();
+  EXPECT_FALSE(HasFailure()) << "Failed on type ArenaBitMap";
+}
